VB-3827 - Move enum into dedicated enum package
