+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonercontactregistry/integration/GetVisitorClosedRestrictionStatusTest.kt
@@ -3,11 +3,11 @@ package uk.gov.justice.digital.hmpps.prisonercontactregistry.integration
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.mock.mockito.SpyBean
-import uk.gov.justice.digital.hmpps.prisonercontactregistry.Restriction
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.client.PrisonApiClient
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.ContactDto
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.ContactsDto
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.RestrictionDto
+import uk.gov.justice.digital.hmpps.prisonercontactregistry.enum.Restriction
 import java.time.LocalDate
 
 @Suppress("ClassName")
